fix(dto): use correct interface coming from backend
